add accuracy measurement to examples and separate qringbuffer logic from connector
