+++ lmcache/integration/vllm/lmcache_mp_connector.py
@@ -3,7 +3,7 @@
 # Standard
 from collections.abc import Iterable
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Literal, Mapping
+from typing import TYPE_CHECKING, Any, Literal
 import enum
 import math
 import sys
@@ -15,7 +15,6 @@ from vllm.distributed.kv_transfer.kv_connector.v1.base import (
     KVConnectorMetadata,
     KVConnectorRole,
 )
-from vllm.v1.kv_cache_interface import AttentionSpec
 
 try:
     # Third Party
@@ -102,6 +101,17 @@ if TYPE_CHECKING:
     from vllm.v1.core.kv_cache_manager import KVCacheBlocks
     from vllm.v1.kv_cache_interface import KVCacheConfig
     from vllm.v1.request import Request
+
+    # First Party
+    from lmcache.sdk.qringbuffer import _QStepState, save_q_layer, setup_q_ring
+else:
+    try:
+        # First Party
+        from lmcache.sdk.qringbuffer import _QStepState, save_q_layer, setup_q_ring
+    except ImportError:
+        _QStepState = None
+        setup_q_ring = None
+        save_q_layer = None
 
 logger = lmcache_init_logger(__name__)
 
@@ -571,84 +581,6 @@ def _ensure_zmq_scheme(server_url: str) -> str:
     return f"tcp://{server_url}"
 
 
-def get_tensor(
-    args: dict[str, Any],
-    arg_names: list[str],
-) -> torch.Tensor | None:
-    """Return the first matching tensor argument, or None.
-
-    Args:
-        args: A dictionary of argument names to values.
-        arg_names: A list of argument names to search for.
-
-    Returns:
-        The first matching tensor argument, None if none found.
-    """
-    for name in arg_names:
-        if name in args:
-            return args[name]
-    return None
-
-
-@dataclass
-class _QLayerStore:
-    """One request's query tensor store for the current forward step.
-
-    Attributes:
-        request_id: The request ID of the stored query tensor.
-        op: The store op whose ``[start, end)`` token range is stored.
-        ring_block_ids: Ring blocks (in token order) holding this op's query.
-        cache_salt: Per-user isolation salt for the query key.
-    """
-
-    request_id: str
-    op: "LoadStoreOp"
-    ring_block_ids: list[int]
-    cache_salt: str
-
-
-@dataclass
-class _QStepState:
-    """Plan to store query tensors for the current forward step, built on the
-    first layer's save_kv_layer call and reused by every layer in the step.
-
-    Attributes:
-        ring_slots: ``int64`` tensor of shape ``[num_tokens]`` mapping each
-            token to ring slot ID, or ``-1`` to drop the row.
-        stores: one ``_QLayerStore`` per store request in the step.
-    """
-
-    ring_slots: torch.Tensor
-    stores: list[_QLayerStore]
-
-
-def attention_layer_names_from_vllm(
-    kv_cache_config: Any,
-    kv_caches: Mapping[str, Any],
-) -> list[str]:
-    """Return the attention layers' names.
-
-    Args:
-        kv_cache_config: vLLM ``KVCacheConfig``, or ``None``.
-        kv_caches: Registered tensors keyed by layer name.
-
-    Returns:
-        Attention layer names, ordered as in ``kv_caches``.
-    """
-    vllm_groups = (
-        getattr(kv_cache_config, "kv_cache_groups", ()) or ()
-        if kv_cache_config is not None
-        else ()
-    )
-    if not vllm_groups:
-        return list(kv_caches.keys())
-    attention_names: set[str] = set()
-    for group in vllm_groups:
-        if isinstance(group.kv_cache_spec, AttentionSpec):
-            attention_names.update(group.layer_names)
-    return [name for name in kv_caches.keys() if name in attention_names]
-
-
 class LMCacheMPConnector(KVConnectorBase_V1, SupportsHMA):
     """
     The connector for LMCache multi-process mode.
@@ -874,71 +806,10 @@ class LMCacheMPConnector(KVConnectorBase_V1, SupportsHMA):
             kv_caches, engine_group_infos=engine_group_infos
         )
         if self.transfer_intermediate_tensors:
-            self._register_q_ring(kv_caches)
-        return
-
-    def _register_q_ring(self, kv_caches: dict[str, torch.Tensor]) -> None:
-        """Allocate and register the paged-Q ring for the attention layers.
-        Builds the mapping between layer names and ring indices (for attention layers).
-        Calls the register_q_ring method of the worker adapter to allocate the ring.
-
-        Args:
-            kv_caches: Registered KV caches.
-        """
-        if not kv_caches:
-            logger.warning("No KV caches registered; skipping Q ring setup.")
-            return
-
-        kv_cache_config = getattr(self, "_kv_cache_config", None)
-        attn_layer_names = attention_layer_names_from_vllm(kv_cache_config, kv_caches)
-        if not attn_layer_names:
-            logger.warning("No attention layers registered. Skipping Q ring setup.")
-            return
-        num_layers = len(attn_layer_names)
-
-        self._q_layer_index = {}
-        for i, name in enumerate(attn_layer_names):
-            self._q_layer_index[name] = i
-
-        model_config = self._vllm_config.model_config
-        parallel_config = self._vllm_config.parallel_config
-        num_q_heads = model_config.get_num_attention_heads(parallel_config)
-        head_size = model_config.get_head_size()
-
-        raw_dtype = model_config.dtype
-        dtype = (
-            raw_dtype
-            if isinstance(raw_dtype, torch.dtype)
-            else getattr(torch, str(raw_dtype))
-        )
-        device = next(iter(kv_caches.values())).device
-        block_size = self._vllm_config.cache_config.block_size
-
-        cfg = self._vllm_config.kv_transfer_config
-        if cfg is None:
-            raise ValueError("kv_transfer_config is None; cannot register Q ring")
-
-        explicit = cfg.get_from_extra_config("lmcache.q.ring_blocks", None)
-        if explicit is not None:
-            num_ring_blocks = max(1, int(explicit))
-        else:
-            depth = max(1, int(cfg.get_from_extra_config("lmcache.q.ring_depth", 2)))
-            max_batched = (
-                getattr(
-                    self._vllm_config.scheduler_config, "max_num_batched_tokens", None
-                )
-                or 8192
+            self._q_layer_index = setup_q_ring(
+                kv_caches, kv_cache_config, self._vllm_config, self.worker_adapter
             )
-            num_ring_blocks = max(1, math.ceil(max_batched / block_size) * depth)
-
-        self.worker_adapter.register_q_ring(
-            num_layers=num_layers,
-            num_q_heads=num_q_heads,
-            head_size=head_size,
-            dtype=dtype,
-            num_ring_blocks=num_ring_blocks,
-            device=device,
-        )
+        return
 
     def start_load_kv(self, forward_context: "ForwardContext", **kwargs: Any) -> None:
         """
@@ -1016,135 +887,17 @@ class LMCacheMPConnector(KVConnectorBase_V1, SupportsHMA):
             attn_metadata (AttentionMetadata): the attention metadata.
             **kwargs: additional arguments for the save operation.
         """
-        if (
-            not self.transfer_intermediate_tensors
-            or not self.worker_adapter.is_kv_writer
-            or self._q_step_disabled
-        ):
-            return
-
-        intermediate_tensors = kwargs.get("intermediate_tensors", None)
-        if intermediate_tensors is None:
-            return
-        query = get_tensor(intermediate_tensors, ["q", "query"])
-        if query is None:
-            return
-
-        layer_index = self._q_layer_index.get(layer_name)
-        if layer_index is None:
-            return
-
-        if self._q_step_state is None:
-            self._q_step_state = self._build_q_step_state(query)
-            if self._q_step_state is None:
-                self._q_step_disabled = True
-                return
-
-        self.worker_adapter.scatter_q_layer(
-            layer_index, query, self._q_step_state.ring_slots
+        self._q_step_state, self._q_step_disabled = save_q_layer(
+            layer_name,
+            self.transfer_intermediate_tensors,
+            self.worker_adapter,
+            self._q_step_disabled,
+            self._q_layer_index,
+            self._q_step_state,
+            self._get_connector_metadata(),
+            **kwargs,
         )
-
-    def _build_q_step_state(self, query: torch.Tensor) -> "_QStepState | None":
-        """Build the per-step query store plan for the current forward step.
-        Done only on the first layer's save_kv_layer call of the step.
-
-        Args:
-            query: The query tensor of shape [num_tokens, num_q_heads, head_size].
-
-        Returns:
-            _QStepState: the query capture plan (containing ring slot mapping
-                         and per-request store ops).
-            None if not storing any query tensors.
-        """
-        metadata = self._get_connector_metadata()
-        assert isinstance(metadata, LMCacheMPConnectorMetadata)
-
-        if not (all(meta.direction == "STORE" for meta in metadata.requests)):
-            logger.warning(
-                "Not all requests are STORE. Skip query for requests %s...",
-                [meta.request_id for meta in metadata.requests],
-            )
-            return None
-
-        q_ring = self.worker_adapter.q_ring
-        if q_ring is None:
-            return None
-        block_size = q_ring.block_size
-        num_query_rows = query.shape[0]
-        ring_slots = torch.full(
-            (num_query_rows,), -1, dtype=torch.int64, device=query.device
-        )
-
-        stores: list[_QLayerStore] = []
-        row_cursor = 0
-        for meta in metadata.requests:
-            op = meta.op
-            num_tokens = op.end - op.start
-            if num_tokens <= 0:
-                continue
-            if meta.direction != "STORE":
-                row_cursor += num_tokens
-                continue
-
-            if row_cursor + num_tokens > num_query_rows:
-                logger.warning(
-                    "Skip query for request %s: query rows (%d) < %d tokens",
-                    meta.request_id,
-                    num_query_rows - row_cursor,
-                    num_tokens,
-                )
-                self._free_partial_ring_blocks(stores)
-                return None
-
-            if num_tokens % block_size != 0:
-                logger.warning(
-                    "Skip query for %s: tokens (%d) undivisible by block size",
-                    meta.request_id,
-                    num_tokens,
-                )
-                self._free_partial_ring_blocks(stores)
-                return None
-
-            n_blocks = num_tokens // block_size
-            ring_blocks = q_ring.allocate(n_blocks)
-            if ring_blocks is None:
-                logger.warning(
-                    "Skip query for request %s: need %d Q blocks, %d free",
-                    meta.request_id,
-                    n_blocks,
-                    q_ring.num_free_blocks(),
-                )
-                self._free_partial_ring_blocks(stores)
-                return None
-
-            block_tensor = torch.tensor(
-                ring_blocks, dtype=torch.int64, device=query.device
-            )
-            rows = torch.arange(num_tokens, device=query.device)
-            slots = block_tensor[rows // block_size] * block_size + (rows % block_size)
-            ring_slots[row_cursor : row_cursor + num_tokens] = slots
-            stores.append(
-                _QLayerStore(
-                    request_id=meta.request_id,
-                    op=op,
-                    ring_block_ids=ring_blocks,
-                    cache_salt=meta.cache_salt,
-                )
-            )
-            row_cursor += num_tokens
-
-        if not stores:
-            return None
-
-        return _QStepState(ring_slots=ring_slots, stores=stores)
-
-    def _free_partial_ring_blocks(self, stores: list[_QLayerStore]) -> None:
-        """Free allocated ring blocks when a full allocation is unsuccessful"""
-        q_ring = self.worker_adapter.q_ring
-        if q_ring is None:
-            return
-        for store in stores:
-            q_ring.free(store.ring_block_ids)
+        return
 
     def wait_for_save(self):
         """
