+++ lmcache/sdk/__init__.py
@@ -3,5 +3,41 @@
 
 # First Party
 from lmcache.sdk import batch, context, kvcache, qcache, stream
+from lmcache.sdk.context import LMCacheSDKCacheKind
 
-__all__ = ["batch", "context", "kvcache", "qcache", "stream"]
+__all__ = ["batch", "context", "kvcache", "qcache", "stream", "LMCacheSDKacheKind"]
+
+
+def connect(
+    kind: LMCacheSDKCacheKind,
+    url: str,
+    http_url: str,
+    model_name: str,
+    timeout: float = 60.0,
+) -> context.LMCacheSDKContext:
+    """Connect to the LMCache server and return a context for the given kind.
+    Calls ``kvcache.connect`` or ``qcache.connect`` based on ``kind``.
+
+    Args:
+        kind: Which cache to connect to (``LMCacheSDKCacheKind.KV`` or
+            ``LMCacheSDKCacheKind.QUERY``).
+        url: ZMQ endpoint URL for the LMCache message queue.
+        http_url: HTTP endpoint URL for fetching server configuration.
+        model_name: Model name used by the running LMCache server instance.
+        timeout: Timeout in seconds for blocking MQ calls. Defaults to 60.
+
+    Returns:
+        An initialized LMCacheSDKContext for the requested kind.
+
+    Raises:
+        ValueError: If ``kind`` is not a supported cache kind.
+    """
+    if kind == LMCacheSDKCacheKind.KV:
+        return kvcache.connect(
+            url=url, http_url=http_url, model_name=model_name, timeout=timeout
+        )
+    if kind == LMCacheSDKCacheKind.QUERY:
+        return qcache.connect(
+            url=url, http_url=http_url, model_name=model_name, timeout=timeout
+        )
+    raise ValueError(f"unsupported cache kind: {kind}")
